new
PDS
The text of the hover effect should be normal in terms of upper and lower case letter. At the moment it’s only upper case.

Publish
Change the text within the search field to “Search for a Project”

Change the tile colour of “Publish new Project” to blue.

Add a hover effect with glow around the edges. This effect should remain when moving the mouse down to the table below. That way it should always be clear which project the below table relates to.

Make the hover table narrower and centralise it underneath the tile that is selected. Sketch below

Scope Business Lounge
The Action Required section should be changed from a dropdown to a fixed element.
Clicking on “Deal with it now” leads to 6. Project Dashboard and opens the appropriate pop-up to either answer a query or to give feedback.